# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API for Studio",
    "version": "1.0.0",
    "description": "Ендпоінти для роботи з авторизацією та управлінням контенту в системі Studio"
  },
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Отримати поточного користувача",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "example": "Bearer <token>" }
          }
        ],
        "responses": {
          "200": {
            "description": "Поточний користувач",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Невалідний або відсутній токен",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "John Doe" },
          "email": { "type": "string", "example": "john@example.com" },
          "role": { "type": "string", "enum": ["USER", "EDITOR", "ADMIN"], "example": "USER" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "email", "role", "createdAt", "updatedAt"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Error message" },
          "statusCode": { "type": "integer", "example": 400 }
        }
      },
...
    }
  }
}